pyproject.toml
[tool.poetry]
name = "plora-swarm"
version = "0.2.1"
description = "Prototype v2 of Plasmid LoRA Swarm"
authors = ["Artem Pitertsev <artem.pitertsev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "plora" }, { include = "swarm" }]

[tool.poetry.dependencies]
python = "3.13.*"
torch = "2.8.0"
transformers = "4.56.1"
peft = "0.12.0"
cryptography = "42.0.7"
pydantic = "2.11.7"
sacrebleu = "2.4.0"
grpcio = "*"
grpcio-tools = "*"
pyyaml = "*"
safetensors = "*"
datasets = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "8.2.0"
ruff = "0.4.3"
black = "24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

